alejandro/requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.1
bcrypt>=4.1.2
python-docx>=1.1.0
flask>=3.0.0
